permissions feature

+++ controllers/admin/tour.controller.js
@@ -119,6 +119,14 @@ module.exports.create = async (req, res) => {
 };
 
 module.exports.createPost = async (req, res) => {
+  if (!req.permissions.includes("tour-create")) {
+    res.json({
+      code: "error",
+      message: "Không có quyền!",
+    });
+    return;
+  }
+
   if (req.body.position) {
     req.body.position = parseInt(req.body.position);
   } else {
@@ -203,6 +211,14 @@ module.exports.edit = async (req, res) => {
 
 module.exports.editPatch = async (req, res) => {
   try {
+    if (!req.permissions.includes("tour-edit")) {
+      res.json({
+        code: "error",
+        message: "Không có quyền!",
+      });
+      return;
+    }
+
     const { id } = req.params;
 
     if (req.body.position) {
@@ -277,6 +293,14 @@ module.exports.editPatch = async (req, res) => {
 
 module.exports.deletePatch = async (req, res) => {
   try {
+    if (!req.permissions.includes("tour-delete")) {
+      res.json({
+        code: "error",
+        message: "Không có quyền!",
+      });
+      return;
+    }
+
     const { id } = req.params;
 
     await Tour.updateOne(
@@ -398,6 +422,14 @@ module.exports.trash = async (req, res) => {
 
 module.exports.undoPatch = async (req, res) => {
   try {
+    if (!req.permissions.includes("tour-trash")) {
+      res.json({
+        code: "error",
+        message: "Không có quyền!",
+      });
+      return;
+    }
+
     const { id } = req.params;
 
     await Tour.updateOne(
@@ -423,6 +455,14 @@ module.exports.undoPatch = async (req, res) => {
 
 module.exports.destroyDelete = async (req, res) => {
   try {
+    if (!req.permissions.includes("tour-trash")) {
+      res.json({
+        code: "error",
+        message: "Không có quyền!",
+      });
+      return;
+    }
+
     const { id } = req.params;
 
     await Tour.deleteOne({
@@ -448,6 +488,14 @@ module.exports.changeMultiPatch = async (req, res) => {
     switch (option) {
       case "active":
       case "inactive":
+        if (!req.permissions.includes("tour-edit")) {
+          res.json({
+            code: "error",
+            message: "Không có quyền!",
+          });
+          return;
+        }
+
         await Tour.updateMany(
           {
             _id: { $in: ids },
@@ -463,6 +511,14 @@ module.exports.changeMultiPatch = async (req, res) => {
         break;
 
       case "undo":
+        if (!req.permissions.includes("tour-trash")) {
+          res.json({
+            code: "error",
+            message: "Không có quyền!",
+          });
+          return;
+        }
+
         await Tour.updateMany(
           {
             _id: { $in: ids },
@@ -478,6 +534,14 @@ module.exports.changeMultiPatch = async (req, res) => {
         break;
 
       case "delete":
+        if (!req.permissions.includes("tour-delete")) {
+          res.json({
+            code: "error",
+            message: "Không có quyền!",
+          });
+          return;
+        }
+
         await Tour.updateMany(
           {
             _id: { $in: ids },
@@ -495,11 +559,17 @@ module.exports.changeMultiPatch = async (req, res) => {
         break;
 
       case "destroy":
-        await Tour.deleteMany(
-          {
-            _id: { $in: ids },
-          },
-        );
+        if (!req.permissions.includes("tour-trash")) {
+          res.json({
+            code: "error",
+            message: "Không có quyền!",
+          });
+          return;
+        }
+
+        await Tour.deleteMany({
+          _id: { $in: ids },
+        });
         res.json({
           code: "success",
           message: "Đã xóa vĩnh viễn!!",
